wire in postings list headers
Right now these simple add a count in addition to an offset within
the postings list.

In the future, this change will enable:
a) quick results for counts of simple (single-term or single-label) queries,
   which actually is a common use case
b) a much more space-efficient postings structure; and
c) much quicker, skiplist-style queries for phrases (which right now are
   pretty slow).

Bump the segment version number to account for this.